+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This list is not currently intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
+### v3.236.0
+- [http4k-testing-servirtium] Don't pass recording handler into non-test methods as a resolved parameter.
+
 ### v3.235.0
 - [http4k-testing-chaos] [Break/Rename] ChaosEngine is now exposed when configuring API. Renamed `withChaosEngine()` to `withChaosApi()`, replaced `toggle()` and `update()` with `enable()/disable()`